backend/backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==6.2.1
certifi==2024.2.2
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.0.0
Flask-Cors==4.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.8.0
pymongo==4.6.0
pyotp==2.9.0
pyparsing==3.2.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
qrcode==7.4.2
reportlab==4.0.7
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
